* Blessings issue better be fixed now, because I can't see any other reason why it wouldn't work.
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/0.2@298 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ luascript.cpp
@@ -7050,7 +7050,7 @@ int32_t LuaScriptInterface::luaGetPlayerBlessing(lua_State* L)
 
 int32_t LuaScriptInterface::luaDoPlayerAddBlessing(lua_State* L)
 {
-	int32_t blessing = popNumber(L);
+	int16_t blessing = popNumber(L);
 	uint32_t cid = popNumber(L);
 	ScriptEnviroment* env = getScriptEnv();
 	Player* player = env->getPlayerByUID(cid);
